add backends query, move stuff to folders
