+++ apps/studio/components/interfaces/DataWarehouse/WarehouseCollectionDetail.tsx
@@ -7,9 +7,9 @@ import ProductEmptyState from 'components/to-be-cleaned/ProductEmptyState'
 import AlertError from 'components/ui/AlertError'
 import LoadingOpacity from 'components/ui/LoadingOpacity'
 import ShimmerLine from 'components/ui/ShimmerLine'
-import { useWarehouseAccessTokensQuery } from 'data/analytics/warehouse-access-tokens-query'
-import { useWarehouseCollectionsQuery } from 'data/analytics/warehouse-collections-query'
-import { useWarehouseQueryQuery } from 'data/analytics/warehouse-query'
+import { useWarehouseAccessTokensQuery } from 'data/analytics/warehouse/warehouse-access-tokens-query'
+import { useWarehouseCollectionsQuery } from 'data/analytics/warehouse/warehouse-collections-query'
+import { useWarehouseQueryQuery } from 'data/analytics/warehouse/warehouse-query'
 import { useFlag } from 'hooks/ui/useFlag'
 import { Button } from 'ui'
 import LogTable from '../Settings/Logs/LogTable'
