+++ apps/studio/components/interfaces/DataWarehouse/WarehouseMenuItem.tsx
@@ -10,8 +10,8 @@ import { z } from 'zod'
 
 import { FormMessage } from '@ui/components/shadcn/ui/form'
 import { useParams } from 'common'
-import { useDeleteCollectionMutation } from 'data/analytics/warehouse-collections-delete-mutation'
-import { useUpdateCollection } from 'data/analytics/warehouse-collections-update-mutation'
+import { useDeleteCollectionMutation } from 'data/analytics/warehouse/warehouse-collections-delete-mutation'
+import { useUpdateCollection } from 'data/analytics/warehouse/warehouse-collections-update-mutation'
 import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
 import {
   Button,
