+++ apps/studio/components/layouts/LogsLayout/LogsLayout.tsx
@@ -7,8 +7,8 @@ import { CreateWarehouseCollectionModal } from 'components/interfaces/DataWareho
 import { WarehouseMenuItem } from 'components/interfaces/DataWarehouse/WarehouseMenuItem'
 import NoPermission from 'components/ui/NoPermission'
 import { ProductMenu } from 'components/ui/ProductMenu'
-import { useWarehouseCollectionsQuery } from 'data/analytics/warehouse-collections-query'
-import { useWarehouseTenantQuery } from 'data/analytics/warehouse-tenant-query'
+import { useWarehouseCollectionsQuery } from 'data/analytics/warehouse/warehouse-collections-query'
+import { useWarehouseTenantQuery } from 'data/analytics/warehouse/warehouse-tenant-query'
 import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
 import { useIsFeatureEnabled } from 'hooks/misc/useIsFeatureEnabled'
 import { useSelectedProject } from 'hooks/misc/useSelectedProject'
