+++ apps/studio/data/analytics/keys.ts
@@ -1,4 +1,7 @@
 export const analyticsKeys = {
+  // Backends
+  backends: (projectRef: string) => ['projects', projectRef, 'backends'] as const,
+
   // warehouse
   warehouseQuery: (projectRef: string, query: string) =>
     ['projects', projectRef, 'warehouse', 'query', query] as const,
