+++ apps/studio/data/analytics/warehouse/warehouse-access-tokens-delete-mutation.ts
@@ -2,7 +2,7 @@ import { useMutation, UseMutationOptions, useQueryClient } from '@tanstack/react
 import toast from 'react-hot-toast'
 import { handleError, del } from 'data/fetchers'
 import type { ResponseError } from 'types'
-import { analyticsKeys } from './keys'
+import { analyticsKeys } from '../keys'
 
 export type DeleteAccessTokenArgs = {
   projectRef: string
