+++ apps/studio/data/analytics/warehouse/warehouse-query.ts
@@ -1,6 +1,6 @@
 import { useQuery, UseQueryOptions } from '@tanstack/react-query'
 import { get, handleError } from 'data/fetchers'
-import { analyticsKeys } from './keys'
+import { analyticsKeys } from '../keys'
 import { ResponseError } from 'types'
 
 export type WarehouseQueryVariables = {
