+++ apps/studio/data/analytics/warehouse/warehouse-tenant-query.ts
@@ -1,5 +1,5 @@
 import { useQuery, UseQueryOptions } from '@tanstack/react-query'
-import { analyticsKeys } from './keys'
+import { analyticsKeys } from '../keys'
 import { get, handleError } from 'data/fetchers'
 import { ResponseError } from 'types'
